\name{arsenate}
\alias{arsenate}
\docType{data}
\title{Comparison of two assays for arsenate.}
\description{Arsenate(V) ion in natural river waters, as determined
by two assay methods.
}
\usage{data(arsenate)}
\format{
  A data frame with 30 observations on the following 4 variables.
  \describe{
    \item{\code{aas}}{micrograms/liter, by continuous selective reduction and
    atomic absorbtion spectectromitry}
    \item{\code{se.aas}}{estimated standard error of the result}
    \item{\code{aes}}{micrograms/liter, by non-selective reduction, 
    cold trapping, and atomic emission spectroscopy}
    \item{\code{se.aes}}{estimated standard error of the result}
  }
}
\source{ 
The data is found in BD Ripley and M Thompson, Regression techniques 
for the detection of analytical bias, Analyst 112:377-383, 1987.
}
\keyword{datasets}
